+++ examples/truetype.rb
@@ -17,7 +17,6 @@
 require 'hexapdf'
 
 doc = HexaPDF::Document.new
-doc.config['font.on_missing_glyph'] = ->(_, f) { f.missing_glyph_id }
 doc.config['font.map'] = {
   'myfont' => {none: ARGV.shift || File.join(__dir__, '../test/data/fonts/Ubuntu-Title.ttf')},
 }
@@ -26,7 +25,7 @@ wrapper = doc.fonts.load('myfont')
 max_gid = wrapper.wrapped_font[:maxp].num_glyphs
 
 255.times do |page|
-  break unless page * 256 < wrapper.wrapped_font[:maxp].num_glyphs
+  break unless page * 256 < max_gid
   canvas = doc.pages.add.canvas
   canvas.font("Helvetica", size: 10)
   canvas.text("Font: #{wrapper.wrapped_font.full_name}", at: [50, 825])
@@ -37,7 +36,7 @@ max_gid = wrapper.wrapped_font[:maxp].num_glyphs
     canvas.show_glyphs((0..15).map do |i|
       gid = page * 256 + y * 16 + i
       glyph = wrapper.glyph(gid)
-      gid > max_gid ? [] : [glyph, -(2000 - glyph.width)]
+      gid >= max_gid ? [] : [glyph, -(2000 - glyph.width)]
     end.flatten!)
   end
 end